#193: Move all arguments to package `***.cmd.arg`
