+++ src/main/java/com/github/dgroup/dockertest/cmd/arg/TimeoutPerThread.java
@@ -21,44 +21,48 @@
  * ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
  * OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package com.github.dgroup.dockertest.cmd;
+package com.github.dgroup.dockertest.cmd.arg;
 
+import com.github.dgroup.dockertest.concurrent.Timeout;
+import com.github.dgroup.dockertest.concurrent.TimeoutOf;
 import java.util.List;
+import java.util.concurrent.TimeUnit;
 import org.cactoos.list.ListOf;
 
 /**
- * Command line argument which represents quantity of concurrent threads for
- *  testing procedure.
+ * Each test will be executed in separate thread.
+ * User may specify the timeout(in seconds) for each test in command line args:
+ *  {@code --timeout-per-test 120}.
  *
- * In case if the argument is missing, the default value is 8.
+ * In case if the argument is missing, the default timeout is 5 minutes.
  *
  * @author Yurii Dubinka (yurii.dubinka@gmail.com)
  * @version $Id$
  * @since 1.0
  */
-public final class ConcurrentTreads extends ArgEnvelope<Integer> {
+public final class TimeoutPerThread extends ArgEnvelope<Timeout> {
 
     /**
      * Ctor.
      * @param args Command-line arguments specified by user.
      */
-    public ConcurrentTreads(final String... args) {
+    public TimeoutPerThread(final String... args) {
         this(new ListOf<>(args));
     }
 
     /**
      * Ctor.
      * @param args Command-line arguments specified by user.
-     * @checkstyle IndentationCheck (10 lines)
      * @checkstyle MagicNumberCheck (10 lines)
      */
-    public ConcurrentTreads(final List<String> args) {
-        super(
-            () -> new Alternative<>(
-                new Mapped<>(Integer::valueOf, new ArgOf("--threads", args)),
-                8
-            )
-        );
+    public TimeoutPerThread(final List<String> args) {
+        super(() -> new Alternative<Timeout>(
+            new Mapped<>(
+                arg -> new TimeoutOf(arg, TimeUnit.SECONDS),
+                new ArgOf("--timeout-per-test", args)
+            ),
+            () -> new TimeoutOf(5, TimeUnit.MINUTES)
+        ));
     }
 
 }
