+++ src/test/java/com/github/dgroup/dockertest/cmd/YmlFileOfTest.java
@@ -23,6 +23,8 @@
  */
 package com.github.dgroup.dockertest.cmd;
 
+import com.github.dgroup.dockertest.cmd.arg.CmdArgNotFoundException;
+import com.github.dgroup.dockertest.cmd.arg.YmlFileOf;
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
 import org.junit.Test;
