+++ src/test/java/com/github/dgroup/dockertest/concurrent/ConcurrentTest.java
@@ -26,8 +26,8 @@ package com.github.dgroup.dockertest.concurrent;
 import com.github.dgroup.dockertest.Assume;
 import com.github.dgroup.dockertest.DockerWasInstalled;
 import com.github.dgroup.dockertest.cmd.Arg;
-import com.github.dgroup.dockertest.cmd.CmdArgNotFoundException;
-import com.github.dgroup.dockertest.cmd.ContainerName;
+import com.github.dgroup.dockertest.cmd.arg.CmdArgNotFoundException;
+import com.github.dgroup.dockertest.cmd.arg.ContainerName;
 import com.github.dgroup.dockertest.hamcrest.HasItems;
 import com.github.dgroup.dockertest.test.TestingFailedException;
 import com.github.dgroup.dockertest.test.output.std.StdOutput;
